fix: update character component
