+++ package.json
@@ -22,6 +22,7 @@
   },
   "devDependencies": {
     "@antfu/eslint-config": "^2.8.0",
+    "@destyler/image": "^0.0.2",
     "@unocss/reset": "^0.58.5",
     "@vitejs/plugin-vue": "^5.0.3",
     "unplugin-auto-import": "^0.17.5",
