+++ src/components/talent.vue
